+++ src/Dialog/styles.ts
@@ -1,5 +1,6 @@
 import { Dialog } from 'react-aria-components';
 import styled, { createGlobalStyle, keyframes } from 'styled-components';
+import { defaultTheme } from '../defaultTheme';
 
 const dialogAnimation = keyframes`
   from {
@@ -16,8 +17,11 @@ export const StyledDialog = styled(Dialog)`
   outline: none;
   box-shadow: 0 0.5rem 1.5rem rgba(0 0 0 / 0.1);
   border-radius: 0.5rem;
-  background: ${({ theme }) => theme.colors.white};
-  border: 1px solid ${({ theme }) => theme.colors.slate[400]};
+  background: ${({ theme }) =>
+    theme?.colors?.white ?? defaultTheme.colors.white};
+  border: 1px solid
+    ${({ theme }) =>
+      theme?.colors?.slate?.[400] ?? defaultTheme.colors.slate[400]};
   outline: none;
   max-width: max-content;
   width: 100vw;