+++ src/Divider/styles.ts
@@ -1,12 +1,14 @@
 import styled from 'styled-components';
 import { SeparatorProps } from 'react-aria';
+import { defaultTheme } from '../defaultTheme';
 
 export const Root = styled.div<{ orientation: SeparatorProps['orientation'] }>`
   width: ${({ orientation }) => (orientation === 'vertical' ? '1px' : '100%')};
   height: ${({ orientation }) => (orientation === 'vertical' ? 'auto' : '1px')};
   margin: ${({ orientation }) =>
     orientation === 'vertical' ? '0 1rem' : '1rem 0'};
-  background: ${({ theme }) => theme.colors?.slate?.[200]};
+  background: ${({ theme }) =>
+    theme?.colors?.slate?.[200] ?? defaultTheme.colors.slate[200]};
 `;
 
 export default Root;