+++ src/Progressbar/LinearProgressbar/styles.ts
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 import { ProgressBar, Label as AriaLabel } from 'react-aria-components';
 import type { LinearProgressbarProps } from './types';
+import { defaultTheme } from '../../defaultTheme';
 
 export const Root = styled(ProgressBar)<{
   width: LinearProgressbarProps['width'];
@@ -31,7 +32,8 @@ export const Bar = styled.div`
   border-radius: 0.3rem;
   overflow: hidden;
   will-change: transform;
-  background: ${({ theme }) => theme.colors.slate[300]};
+  background: ${({ theme }) =>
+    theme?.colors?.slate?.[300] ?? defaultTheme.colors.slate[300]};
 `;
 
 export const Fill = styled.div<{
@@ -42,7 +44,8 @@ export const Fill = styled.div<{
   transition: ${({ isIndeterminate }) =>
     isIndeterminate ? 'none' : 'width 600ms ease-in-out'};
   height: 100%;
-  background: ${({ theme }) => theme.colors.slate[800]};
+  background: ${({ theme }) =>
+    theme?.colors?.slate?.[800] ?? defaultTheme.colors.slate[800]};
   border-radius: inherit;
   animation: ${({ isIndeterminate }) =>
     isIndeterminate ? 'indeterminate 1.5s infinite ease-in-out' : 'none'};