+++ src/Select/styles.tsx
@@ -5,14 +5,13 @@ import {
   FieldError,
   Input,
   Label,
-  ListBox,
   ListBoxItem,
   Popover,
   Text,
   TextField,
 } from 'react-aria-components';
 import { motion } from 'framer-motion';
-import { ReactNode } from 'react';
+import { defaultTheme } from '../defaultTheme';
 
 export const StyledComboBox = styled(ComboBox)`
   display: flex;
@@ -32,61 +31,91 @@ export const StyledPopover = motion(styled(Popover)`
   border-radius: 0.375rem;
   width: 14rem;
   background: #fff;
-  border: 1px solid ${({ theme }) => theme.colors.slate[300]};
+  border: 1px solid
+    ${({ theme }) =>
+      theme?.colors?.slate?.[300] ?? defaultTheme.colors.slate[300]};
 `);
 
 export const StyledInput = styled(Input).withConfig({
   shouldForwardProp: (prop) => !['isLoading'].includes(prop),
 })<{ isLoading: boolean }>`
-  border: 1px solid ${({ theme }) => theme.colors.slate[300]};
+  border: 1px solid
+    ${({ theme }) =>
+      theme?.colors?.slate?.[300] ?? defaultTheme.colors.slate[300]};
   font-size: 0.9rem;
   padding: 0 0.5rem;
   height: 2.5rem;
   margin: 0;
   border-radius: 0.375rem;
-  background: ${({ theme }) => theme.colors.white};
-  color: ${({ theme }) => theme.colors.slate[900]};
+  background: ${({ theme }) =>
+    theme?.colors?.white ?? defaultTheme.colors.white};
+  color: ${({ theme }) =>
+    theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]};
   width: 100%;
   transition: all ease 200ms;
 
   &[data-hovered] {
-    border: 1px solid ${({ theme }) => theme.colors.slate[400]};
+    border: 1px solid
+      ${({ theme }) =>
+        theme?.colors?.slate?.[400] ?? defaultTheme.colors.slate[400]};
   }
 
   &[data-focused] {
-    border: 1px solid ${({ theme }) => theme.colors.slate[400]};
-    outline: 2px solid ${({ theme }) => theme.colors.slate[200]};
+    border: 1px solid
+      ${({ theme }) =>
+        theme?.colors?.slate?.[400] ?? defaultTheme.colors.slate[400]};
+    outline: 2px solid
+      ${({ theme }) =>
+        theme?.colors?.slate?.[200] ?? defaultTheme.colors.slate[200]};
   }
 
   &[data-disabled] {
     background: ${({ isLoading, theme }) =>
-      isLoading ? theme.colors.white : theme.colors.slate[100]};
-    border: 1px solid ${({ theme }) => theme.colors.slate[300]};
+      isLoading
+        ? theme?.colors?.white ?? defaultTheme.colors.white
+        : theme?.colors?.slate?.[100] ?? defaultTheme.colors.slate[100]};
+    border: 1px solid
+      ${({ theme }) =>
+        theme?.colors?.slate?.[300] ?? defaultTheme.colors.slate[300]};
     color: ${({ isLoading, theme }) =>
-      isLoading ? theme.colors.slate[900] : theme.colors.slate[500]};
+      isLoading
+        ? theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]
+        : theme?.colors?.slate?.[500] ?? defaultTheme.colors.slate[500]};
   }
 
   &[data-invalid] {
-    border: 1px solid ${({ theme }) => theme.colors.red[500]};
-    background: ${({ theme }) => theme.colors.red[100]};
-    color: ${({ theme }) => theme.colors.red[600]};
+    border: 1px solid
+      ${({ theme }) =>
+        theme?.colors?.red?.[500] ?? defaultTheme.colors.red[500]};
+    background: ${({ theme }) =>
+      theme?.colors?.red?.[100] ?? defaultTheme.colors.red[100]};
+    color: ${({ theme }) =>
+      theme?.colors?.red?.[600] ?? defaultTheme.colors.red[600]};
 
     &[data-hovered] {
-      border: 1px solid ${({ theme }) => theme.colors.red[600]};
+      border: 1px solid
+        ${({ theme }) =>
+          theme?.colors?.red?.[600] ?? defaultTheme.colors.red[600]};
     }
 
     &[data-focused] {
-      border: 1px solid ${({ theme }) => theme.colors.red[600]};
-      outline: 2px solid ${({ theme }) => theme.colors.red[200]};
+      border: 1px solid
+        ${({ theme }) =>
+          theme?.colors?.red?.[600] ?? defaultTheme.colors.red[600]};
+      outline: 2px solid
+        ${({ theme }) =>
+          theme?.colors?.red?.[200] ?? defaultTheme.colors.red[200]};
     }
 
     ::placeholder {
-      color: ${({ theme }) => theme.colors.red[500]};
+      color: ${({ theme }) =>
+        theme?.colors?.red?.[500] ?? defaultTheme.colors.red[500]};
     }
   }
 
   ::placeholder {
-    color: ${({ theme }) => theme.colors.slate[400]};
+    color: ${({ theme }) =>
+      theme?.colors?.slate?.[400] ?? defaultTheme.colors.slate[400]};
   }
 `;
 
@@ -104,13 +133,13 @@ export const StyledButtonIcon = styled.svg.withConfig({
   path {
     stroke: ${({ isDisabled, isInvalid, theme }) => {
       if (isInvalid) {
-        return theme.colors.red[500];
+        return theme?.colors?.red?.[500] ?? defaultTheme.colors.red[500];
       }
       if (isDisabled) {
-        return theme.colors.slate[400];
+        return theme?.colors?.slate?.[400] ?? defaultTheme.colors.slate[400];
       }
 
-      return theme.colors.slate[900];
+      return theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900];
     }};
   }
 `;
@@ -134,25 +163,29 @@ export const StyledButton = styled(Button)`
 export const StyledLabel = styled(Label)`
   display: flex;
   font-size: 0.8rem;
-  color: ${({ theme }) => theme.colors.slate[500]};
+  color: ${({ theme }) =>
+    theme?.colors?.slate?.[500] ?? defaultTheme.colors.slate[500]};
 `;
 
 export const StyledTextField = styled(TextField)`
   display: flex;
   flex-direction: column;
   width: 100%;
-  color: ${({ theme }) => theme.colors.slate[900]};
+  color: ${({ theme }) =>
+    theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]};
 `;
 
 export const StyledFieldError = styled(FieldError)`
   display: flex;
-  color: ${({ theme }) => theme.colors.red[500]};
+  color: ${({ theme }) =>
+    theme?.colors?.red?.[500] ?? defaultTheme.colors.red[500]};
   font-size: 0.8rem;
 `;
 
 export const StyledText = styled(Text)`
   display: flex;
-  color: ${({ theme }) => theme.colors.slate[500]};
+  color: ${({ theme }) =>
+    theme?.colors?.slate?.[500] ?? defaultTheme.colors.slate[500]};
   font-size: 0.8rem;
 `;
 
@@ -167,37 +200,47 @@ export const StyledListBoxItem = styled(ListBoxItem)`
   margin-right: 0.4rem;
   cursor: default;
   outline: none;
-  color: ${({ theme }) => theme.colors.slate[900]};
+  color: ${({ theme }) =>
+    theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]};
   transition: all ease 0.25s;
 
   &[data-hovered] {
-    background: ${({ theme }) => theme.colors.slate[100]};
-    color: ${({ theme }) => theme.colors.slate[900]};
+    background: ${({ theme }) =>
+      theme?.colors?.slate?.[100] ?? defaultTheme.colors.slate[100]};
+    color: ${({ theme }) =>
+      theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]};
   }
 
   &[data-focused] {
-    background: ${({ theme }) => theme.colors.slate[100]};
-    color: ${({ theme }) => theme.colors.slate[900]};
+    background: ${({ theme }) =>
+      theme?.colors?.slate?.[100] ?? defaultTheme.colors.slate[100]};
+    color: ${({ theme }) =>
+      theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]};
   }
 
   &[data-pressed] {
-    background: ${({ theme }) => theme.colors.slate[200]};
-    color: ${({ theme }) => theme.colors.slate[900]};
+    background: ${({ theme }) =>
+      theme?.colors?.slate?.[200] ?? defaultTheme.colors.slate[200]};
+    color: ${({ theme }) =>
+      theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]};
   }
 
   &[data-selected] {
     background: none;
-    color: ${({ theme }) => theme.colors.slate[900]};
+    color: ${({ theme }) =>
+      theme?.colors?.slate?.[900] ?? defaultTheme.colors.slate[900]};
     font-weight: 600;
 
     &[data-focused] {
-      background: ${({ theme }) => theme.colors.slate[100]};
+      background: ${({ theme }) =>
+        theme?.colors?.slate?.[100] ?? defaultTheme.colors.slate[100]};
     }
   }
 
   &[data-disabled] {
     background: none;
-    color: ${({ theme }) => theme.colors.slate[500]};
+    color: ${({ theme }) =>
+      theme?.colors?.slate?.[500] ?? defaultTheme.colors.slate[500]};
   }
 `;