+++ src/Tabs/styles.tsx
@@ -2,6 +2,7 @@ import styled, { css } from 'styled-components';
 import { Tabs, TabList, Tab, TabPanel, TabsProps } from 'react-aria-components';
 import { motion } from 'framer-motion';
 import type { TabListProps, TabItem, TabProps } from './types';
+import { defaultTheme } from '../defaultTheme';
 
 export const StyledTabs = styled(Tabs)`
   width: 100%;
@@ -32,13 +33,15 @@ export const StyledTabList = styled(TabList<TabItem>)<TabListProps>`
     switch (variant) {
       case 'solid':
         return css`
-          background-color: ${theme.colors.slate[200]};
+          background-color: ${theme?.colors?.slate?.[200] ??
+          defaultTheme.colors.slate[200]};
           border-radius: 0.75rem;
           padding: 0.25rem;
         `;
       case 'outline':
         return css`
-          border: 2px solid ${theme.colors.slate[200]};
+          border: 2px solid
+            ${theme?.colors?.slate?.[200] ?? defaultTheme.colors.slate[200]};
           border-radius: 0.75rem;
           padding: 0.25rem;
         `;
@@ -53,7 +56,8 @@ export const StyledTabList = styled(TabList<TabItem>)<TabListProps>`
               bottom: 0;
               content: '';
               width: 1px;
-              background-color: ${theme.colors.slate[200]};
+              background-color: ${theme?.colors?.slate?.[200] ??
+              defaultTheme.colors.slate[200]};
             }
           }
           &[data-orientation='horizontal'] {
@@ -65,7 +69,8 @@ export const StyledTabList = styled(TabList<TabItem>)<TabListProps>`
               right: 0;
               content: '';
               height: 1px;
-              background-color: ${theme.colors.slate[200]};
+              background-color: ${theme?.colors?.slate?.[200] ??
+              defaultTheme.colors.slate[200]};
             }
           }
         `;
@@ -111,9 +116,9 @@ export const StyledTab = styled(Tab)<
       case 'underline':
       case 'enclosed':
       case 'outline':
-        return theme.colors.slate[800];
+        return theme?.colors?.slate?.[800] ?? defaultTheme.colors.slate[800];
       case 'solid':
-        return theme.colors.slate[400];
+        return theme?.colors?.slate?.[400] ?? defaultTheme.colors.slate[400];
       default:
         return null;
     }
@@ -124,10 +129,10 @@ export const StyledTab = styled(Tab)<
       switch (variant) {
         case 'underline':
         case 'outline':
-          return theme.colors.slate[800];
+          return theme?.colors?.slate?.[800] ?? defaultTheme.colors.slate[800];
         case 'solid':
         case 'enclosed':
-          return theme.colors.slate[100];
+          return theme?.colors?.slate?.[100] ?? defaultTheme.colors.slate[100];
         default:
           return null;
       }
@@ -142,7 +147,8 @@ export const StyledTab = styled(Tab)<
   }
 
   &[data-hovered]:not([data-selected]) {
-    color: ${({ theme }) => theme.colors.slate[500]};
+    color: ${({ theme }) =>
+      theme?.colors?.slate?.[500] ?? defaultTheme.colors.slate[500]};
   }
 
   &[data-disabled] {
@@ -162,13 +168,15 @@ export const TabOverlay = styled(motion.span).attrs({
         return css`
           inset: 0px;
           border-radius: 0.5rem;
-          background-color: ${theme.colors.slate[600]};
+          background-color: ${theme?.colors?.slate?.[600] ??
+          defaultTheme.colors.slate[600]};
           mix-blend-mode: color;
         `;
       case 'outline':
         return css`
           inset: 0px;
-          border: 2px solid ${theme.colors.slate[800]};
+          border: 2px solid
+            ${theme?.colors?.slate?.[800] ?? defaultTheme.colors.slate[800]};
           border-radius: 0.5rem;
         `;
       case 'underline':
@@ -179,7 +187,8 @@ export const TabOverlay = styled(motion.span).attrs({
             width: 100%;
             height: 1px;
             border-radius: 0.5rem;
-            background-color: ${theme.colors.slate[800]};
+            background-color: ${theme?.colors?.slate?.[800] ??
+            defaultTheme.colors.slate[800]};
             mix-blend-mode: color;
           `;
         }
@@ -190,17 +199,20 @@ export const TabOverlay = styled(motion.span).attrs({
           width: 1px;
           height: 100%;
           border-radius: 0.5rem;
-          background-color: ${theme.colors.slate[800]};
+          background-color: ${theme?.colors?.slate?.[800] ??
+          defaultTheme.colors.slate[800]};
           mix-blend-mode: color;
         `;
       case 'enclosed':
         return css`
           inset: 0px;
-          border: 1px solid ${theme.colors.slate[800]};
+          border: 1px solid
+            ${theme?.colors?.slate?.[800] ?? defaultTheme.colors.slate[800]};
           margin-bottom: -1px;
           border-bottom: none;
           border-radius: 0.5rem 0.5rem 0 0;
-          background-color: ${theme.colors.white};
+          background-color: ${theme?.colors?.white ??
+          defaultTheme.colors.white};
           mix-blend-mode: difference;
         `;
       case 'base':